requirements.txt
mesa
ipykernel
plotly
nbformat
scipy
numba
black
pyinstrument
pytest
streamlit
pyogrio
geopandas
fiona
seaborn
pydantic
streamlit-mermaid
dill